requirements.txt
python-dotenv
fastapi
sqlalchemy
uvicorn
pymysql